(1) Fixing line endings; (2) Removed ipynb checkpoints being committed (3) Fixing bugs in relu_backward functoin (4) Added CNN (5) Fixed bug in FC Network

+++ assignment2/cs231n/solver.py
@@ -249,7 +249,7 @@ class Solver(object):
     """
     num_train = self.X_train.shape[0]
     iterations_per_epoch = max(num_train / self.batch_size, 1)
-    num_iterations = self.num_epochs * iterations_per_epoch
+    num_iterations = int(self.num_epochs * iterations_per_epoch)
 
     for t in range(num_iterations):
       self._step()
